backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
google-cloud-aiplatform
langchain
langchain-google-vertexai